requirements.txt
fastapi
uvicorn
twilio
google-genai
httpx
python-multipart
python-dotenv